LIB:  Consensus sorting algorithms & new parsers to bitmapped_bytes
Added classes and functions for distrubed algorithms (so far just a couple of
consensus sorting algorithms).

New parsers mostly for variable length messages.  Concrete classes for
a number of uniform lists of same-type numerical values were implemented.
Also a couple of abstract classes for more generic variable length
data payloads, and a class for parsing ID-value pairs (which will be used
in the consensus algorithm sorting).

Added a publisher for the send_swarm_behavior_data topic to the behavior class
so that inheriting behaviors can publish swarm data messages for transmission
over the network.