PATH_PLANNING: Added an "L1" calculation to intercept calculator
Added a simple L1 calculation to the intercept calculator to improve the
performance of the linear formation calculator.  Also wrote a potential
field intercept calculator.  In testing, it performs similarly to the
other intercept calculator in the linear formation controller, but is not
currently being used.  It may be useful as the basis for a more Boyds-like
flocking implementation at some point.