PATH_PLANNING: Added an independent transit behavior
The new behavior will cause the UAV to transit to a specific geographic
location at the currently ordered altitude.  All UAVs within the subswarm
will transit to the same location, but the transit will not be coordinated
(they all just go directly to the location).

Also changed the swarm search and greedy shooter behaviors so that they
remain "active" when the behavior finishes (they will return to the
standby waypoint).  This will allow for cleaner scripting of "plays".