PATH_PLANNING:  PN Interceptor implementation & PN for greedy_shooter
Initial implementation of the PN Interceptor behavior.  Only works for
a straight run-in scenario, and does not yet incorporate the terminal phase.

Also switched the greedy_shooter pursuit algorithm to use a proportional
navigation function instead of a direct-to homing function.

Also switched the linear formation and landing sequencer consensus algorithm
to use the eager algorithm (vice lazy).